api: unify auth headers
