+++ src/lib/api/houses.ts
@@ -1,11 +1,10 @@
 import { HouseConfig, defaultHouses } from '@/types/inventory';
-import { API_URL, API_KEY } from './common';
+import { API_URL, withAuthHeaders } from './common';
 
 function buildHeaders(contentType?: string) {
   const headers: Record<string, string> = {};
   if (contentType) headers['Content-Type'] = contentType;
-  if (API_KEY) headers['X-API-Key'] = API_KEY;
-  return headers;
+  return withAuthHeaders(headers);
 }
 
 function getAllHouses(): HouseConfig[] {
@@ -100,6 +99,7 @@ export async function deleteHouse(id: string) {
   try {
     const response = await fetch(`${API_URL}/houses/${id}`, {
       method: 'DELETE',
+      headers: buildHeaders(),
     });
     if (!response.ok) throw new Error('Failed to delete house');
     const houses = getAllHouses().map((h) =>
