+++ src/lib/api/items.ts
@@ -8,20 +8,10 @@ import {
 import { testDecorItems } from '@/data/testData';
 import { testBooks } from '@/data/booksTestData';
 import { testMusic } from '@/data/musicTestData';
-import { API_URL, API_KEY } from './common';
+import { API_URL, withAuthHeaders } from './common';
 
 const isApiConfigured = () => {
   return !!API_URL;
-};
-
-const withOptionalAuth = (
-  base: Record<string, string> = {},
-): Record<string, string> => {
-  const headers = { ...base };
-  if (API_KEY) {
-    headers.Authorization = `Bearer ${API_KEY}`;
-  }
-  return headers;
 };
 
 const isTestDataEnabled = () => {
@@ -129,7 +119,7 @@ export async function fetchDecorItems(): Promise<DecorItem[]> {
   
   try {
     const response = await fetch(`${API_URL}/items`, {
-      headers: withOptionalAuth({ 'Content-Type': 'application/json' }),
+      headers: withAuthHeaders({ 'Content-Type': 'application/json' }),
     });
     if (!response.ok) throw new Error('API request failed');
     return await response.json();
@@ -156,7 +146,7 @@ export async function fetchDecorItem(
   
   try {
     const response = await fetch(`${API_URL}/items/${id}`, {
-      headers: withOptionalAuth({ 'Content-Type': 'application/json' }),
+      headers: withAuthHeaders({ 'Content-Type': 'application/json' }),
     });
     if (!response.ok) throw new Error('Failed to fetch item');
     return await response.json();
@@ -179,7 +169,7 @@ export async function fetchBookItems(): Promise<BookItem[]> {
   
   try {
     const response = await fetch(`${API_URL}/books`, {
-      headers: withOptionalAuth({ 'Content-Type': 'application/json' }),
+      headers: withAuthHeaders({ 'Content-Type': 'application/json' }),
     });
     if (!response.ok) throw new Error('API request failed');
     return await response.json();
@@ -202,7 +192,7 @@ export async function fetchMusicItems(): Promise<MusicItem[]> {
   
   try {
     const response = await fetch(`${API_URL}/music`, {
-      headers: withOptionalAuth({ 'Content-Type': 'application/json' }),
+      headers: withAuthHeaders({ 'Content-Type': 'application/json' }),
     });
     if (!response.ok) throw new Error('API request failed');
     return await response.json();
@@ -231,7 +221,7 @@ export async function createDecorItem(
   try {
     const response = await fetch(`${API_URL}/items`, {
       method: 'POST',
-      headers: withOptionalAuth({ 'Content-Type': 'application/json' }),
+      headers: withAuthHeaders({ 'Content-Type': 'application/json' }),
       body: JSON.stringify(input),
     });
     if (!response.ok) throw new Error('Failed to create item');
@@ -270,7 +260,7 @@ export async function updateDecorItem(
   try {
     const response = await fetch(`${API_URL}/items/${id}`, {
       method: 'PUT',
-      headers: withOptionalAuth({ 'Content-Type': 'application/json' }),
+      headers: withAuthHeaders({ 'Content-Type': 'application/json' }),
       body: JSON.stringify(input),
     });
     if (!response.ok) throw new Error('Failed to update item');
@@ -295,7 +285,7 @@ export async function deleteDecorItem(id: number): Promise<void> {
   try {
     const response = await fetch(`${API_URL}/items/${id}`, {
       method: 'DELETE',
-      headers: withOptionalAuth(),
+      headers: withAuthHeaders(),
     });
     if (!response.ok) throw new Error('Failed to delete item');
   } catch (error) {
@@ -330,7 +320,7 @@ export async function restoreDecorItem(
   try {
     const response = await fetch(`${API_URL}/items/${id}/restore`, {
       method: 'POST',
-      headers: withOptionalAuth({ 'Content-Type': 'application/json' }),
+      headers: withAuthHeaders({ 'Content-Type': 'application/json' }),
       body: JSON.stringify(version),
     });
     if (!response.ok) throw new Error('Failed to restore item');
